Updated test codes and repositories. - 업데이트된 엔티티(Long 타입 식별자 등)에 따라 테스트 코드 수정. - 바우처 타입에 따라 생성, 갱신, 제약사항 적용 테스트 코드 분리. - 바우처 서비스에서 정의하는 기능 중 누락된 테스트 코드 작성. - 회원 조회 방식에 따른 테스트 코드 분리. - 바우처 저장 시 created_at 필드를 지정하지 않아 서버와 DB 시간대 차이로 인한 타임스탬프 오류 수정. - local file repository 에서 예외 발생 시 메시지를 로그로 출력하도록 변경.

+++ voucher/src/test/java/com/programmers/voucher/service/relation/CustomerWalletTest.java
@@ -65,7 +65,7 @@ class CustomerWalletTest {
     @DisplayName("Get vouchers of customer")
     void getVouchers() {
         List<Voucher> list = new ArrayList<>(5);
-        for (int i = 0; i < 5; i++) {
+        for (long i = 0; i < 5; i++) {
             list.add(new Voucher(i + 1, "voucherName" + i, new DiscountPolicy(2500, DiscountType.FIXED), LocalDate.now(), 1000 + i));
         }
         long customerId = 1;
